BrokenRecord

BrokenRecord.configure!

BrokenRecord.playback

BrokenRecord.playback — Method

playback(f, path; ignore_headers=[], ignore_query=[])

Run f while either recording to or playing back from the file at path. See configure! for more information on the available keywords.

BrokenRecord.configure! — Method

configure!(; path=nothing, ignore_headers=nothing, ignore_query=nothing)

Set options globally so that you needn't pass keywords to every playback call.

Keywords

path: Path to the directory that contains data files. Any path you pass to playback will be relative to this path.

ignore_headers: Names of headers to remove from requests.

ignore_query: Names of query string parameters to remove from requests.